페이징, csBoard페이징, 리뷰에러
리뷰작성시 point부분들 전부 update되게 진행함, 페이징 오류 확인, csBoard페이징확인, 리뷰수정하기, 삭제하기, 좋아요 에러확인, connection이용하여 product에 삽입할 insert한 reviewNo 를 찾는 부분 구현 중

+++ src/main/webapp/review/delete.jsp
@@ -1,15 +1,26 @@
+<%@page import="vo.UserTable"%>
 <%@page import="vo.Review"%>
 <%@page import="dao.ReviewJdbcDao"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 <%
+	UserTable loginUserInfo = (UserTable)session.getAttribute("LOGIN_USER_INFO");
+
 	int no = Integer.parseInt(request.getParameter("no"));
 	int cpno = Integer.parseInt(request.getParameter("cpno"));
 	int reviewNo = Integer.parseInt(request.getParameter("reviewNo"));
 
+	if(loginUserInfo == null) {
+		response.sendRedirect("../main/loginform.jsp?error=empty");
+		return;
+	}
+	if(reviewNo == 0) {//title이 null이 아니고,title의 값이 비어있으면
+		response.sendRedirect("reviewUserDetail.jsp?no="+no+"&cpno="+cpno+"&error=empty");
+		return;
+	}
 	ReviewJdbcDao reviewDao = ReviewJdbcDao.getInstance();
-	Review review = reviewDao.getReviewById(reviewNo);
+	Review review = reviewDao.getReviewByReviewNo(reviewNo);
 	reviewDao.deleteReviewById(review.getReviewNo());
 	
-	response.sendRedirect("detail.jsp?cpno="+cpno);
+	response.sendRedirect("detail.jsp?cpno="+cpno+"&error=ok");
 
 %>
